avoid double to int warnings with gd_alpha computation (#2275)
git-svn-id: http://svn.osgeo.org/mapserver/trunk@6743 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapresample.c
@@ -532,7 +532,7 @@ msBilinearRasterResampler( imageObj *psSrcImage, colorObj offsite,
                     else
                     {
                         int gd_color;
-                        int gd_alpha = 127 - 127.9 * dfWeightSum;
+                        int gd_alpha = (int) (127 - 127.9 * dfWeightSum);
 
                         gd_alpha = MAX(0,MIN(127,gd_alpha));
                         gd_color = gdTrueColorAlpha(
@@ -796,7 +796,7 @@ msAverageRasterResampler( imageObj *psSrcImage, colorObj offsite,
                     else
                     {
                         int gd_color;
-                        int gd_alpha = 127 - 127.9 * dfAlpha01;
+                        int gd_alpha = (int) (127 - 127.9 * dfAlpha01);
 
                         gd_alpha = MAX(0,MIN(127,gd_alpha));
                         gd_color = gdTrueColorAlpha(
@@ -835,7 +835,7 @@ msAverageRasterResampler( imageObj *psSrcImage, colorObj offsite,
                     else
                     {
                         int gd_color;
-                        int gd_alpha = 127 - 127.9 * dfAlpha01;
+                        int gd_alpha = (int) (127 - 127.9 * dfAlpha01);
 
                         gd_alpha = MAX(0,MIN(127,gd_alpha));
                         gd_color = gdTrueColorAlpha(
